Move storage-related type traits from matrix to storage files
