+++ include/alp/reference/matrix.hpp
@@ -364,96 +364,6 @@ namespace alp {
 
 	} // namespace internal
 
-	// Matrix-related implementation
-
-	namespace internal {
-
-
-		/** Specialization for general matrix */
-		template<>
-		struct determine_poly_factory< structures::General, imf::Id, imf::Id, reference > {
-
-			typedef storage::polynomials::FullFactory<> factory_type;
-		};
-
-		/** Specialization for square matrix */
-		template<>
-		struct determine_poly_factory< structures::Square, imf::Id, imf::Id, reference > {
-
-			typedef storage::polynomials::FullFactory<> factory_type;
-		};
-
-		/** Specialization for orthogonal matrix */
-		template<>
-		struct determine_poly_factory< structures::Orthogonal, imf::Id, imf::Id, reference > {
-
-			typedef storage::polynomials::FullFactory<> factory_type;
-		};
-
-		/** Specialization for upper-triangular matrix */
-		template<>
-		struct determine_poly_factory< structures::UpperTriangular, imf::Id, imf::Id, reference > {
-
-			typedef storage::polynomials::PackedFactory< storage::UPPER, storage::ROW_WISE > factory_type;
-		};
-
-		/** Specialization for lower-triangular matrix */
-		template<>
-		struct determine_poly_factory< structures::LowerTriangular, imf::Id, imf::Id, reference > {
-
-			typedef storage::polynomials::PackedFactory< storage::LOWER, storage::ROW_WISE > factory_type;
-		};
-
-		/** Specialization for symmetric matrix */
-		template<>
-		struct determine_poly_factory< structures::Symmetric, imf::Id, imf::Id, reference > {
-
-			typedef storage::polynomials::PackedFactory< storage::UPPER, storage::ROW_WISE > factory_type;
-		};
-
-		/** Specialization for hermitian matrix */
-		template<>
-		struct determine_poly_factory< structures::Hermitian, imf::Id, imf::Id, reference > {
-
-			typedef storage::polynomials::FullFactory<> factory_type;
-		};
-
-		/** Specialization for symmetric tridiagonal matrix */
-		template<>
-		struct determine_poly_factory< structures::SymmetricTridiagonal, imf::Id, imf::Id, reference > {
-
-			private:
-				using interval = std::tuple_element< 0, structures::SymmetricTridiagonal::band_intervals >::type;
-
-			public:
-				//typedef storage::polynomials::BandFactory< interval, storage::ROW_WISE > factory_type;
-				typedef storage::polynomials::FullFactory<> factory_type;
-		};
-
-		/** Specialization for hermitian tridiagonal matrix */
-		template<>
-		struct determine_poly_factory< structures::HermitianTridiagonal, imf::Id, imf::Id, reference > {
-
-			private:
-				// This will be used in the commented line below once band storage is added.
-				// Added for readability.
-				using interval = std::tuple_element< 0, structures::SymmetricTridiagonal::band_intervals >::type;
-
-			public:
-				//typedef storage::polynomials::BandFactory< interval, storage::ROW_WISE > factory_type;
-				typedef storage::polynomials::FullFactory<> factory_type;
-		};
-
-		/** Specialization for vectors */
-		template< typename Structure >
-		struct determine_poly_factory< Structure, imf::Id, imf::Zero, reference > {
-
-			typedef storage::polynomials::ArrayFactory factory_type;
-		};
-
-	} // namespace internal
-
-
 } // namespace alp
 
 #endif // end ``_H_ALP_REFERENCE_MATRIX''
